added new PrismApplication class
The bootstrapper has been deprecated. A base application base class is
needed in order to handle the various mobile scenarios that will not be
possible with using a bootstrapper.
