+++ Samples/Xamarin/HelloWorld/HelloWorld/HelloWorld/App.cs
@@ -1,33 +1,24 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-
-using Xamarin.Forms;
+﻿using HelloWorld.Views;
+using Prism.Unity;
 
 namespace HelloWorld
 {
-    public class App : Application
+    public class App : PrismApplication
     {
-        public App()
+        protected override void OnInitialized()
         {
-            Bootstrapper bs = new Bootstrapper();
-            bs.Run(this);
+            NavigationService.Navigate("MyMasterDetail/MyNavigationPage/ViewA/MyTabbedPage/ViewB");
         }
 
-        protected override void OnStart()
+        protected override void RegisterTypes()
         {
-            // Handle when your app starts
-        }
-
-        protected override void OnSleep()
-        {
-            // Handle when your app sleeps
-        }
-
-        protected override void OnResume()
-        {
-            // Handle when your app resumes
+            Container.RegisterTypeForNavigation<MainPage>();
+            Container.RegisterTypeForNavigation<MyNavigationPage>();
+            Container.RegisterTypeForNavigation<MyTabbedPage>();
+            Container.RegisterTypeForNavigation<MyMasterDetail>();
+            Container.RegisterTypeForNavigation<ViewA>();
+            Container.RegisterTypeForNavigation<ViewB>();
+            Container.RegisterTypeForNavigation<ViewC>();
         }
     }
 }
