+++ Source/Xamarin/Prism.Forms/ApplicationBase.cs
@@ -1,24 +1,11 @@
 ﻿using Prism.Logging;
 using Prism.Navigation;
-using System;
 using Xamarin.Forms;
 
 namespace Prism
 {
-    /// <summary>
-    /// Base class that provides a basic bootstrapping sequence and hooks that specific implementations can override
-    /// </summary>
-    /// <remarks>
-    /// This class must be overridden to provide application specific configuration.
-    /// </remarks>
-    [Obsolete("Please have your App.cs derive from PrismApplication instead.")]
-    public abstract class Bootstrapper
+    public abstract class PrismApplicationBase : Application
     {
-        /// <summary>
-        /// The current <see cref="Xamarin.Forms.Application"/>
-        /// </summary>
-        protected Application App { get; set; }
-
         /// <summary>
         /// Gets the <see cref="ILoggerFacade"/> for the application.
         /// </summary>
@@ -30,24 +17,19 @@ namespace Prism
         /// </summary>
         protected INavigationService NavigationService { get; set; }
 
-        /// <summary>
-        /// Runs the bootstrapper process.
-        /// </summary>
-        /// <param name="app">The current <see cref="Xamarin.Forms.Application"/></param>
-        public void Run(Application app)
+        public PrismApplicationBase()
         {
-            App = app;
             ConfigureViewModelLocator();
-            Run();
+            Initialize();
         }
 
         /// <summary>
         /// Run the bootstrapper process.
         /// </summary>
-        public abstract void Run();
+        public abstract void Initialize();
 
         /// <summary>
-        /// Create the <see cref="ILoggerFacade" /> used by the bootstrapper.
+        /// Create the <see cref="ILoggerFacade" /> used by the application.
         /// </summary>
         /// <remarks>
         /// The base implementation returns a new TextLogger.
@@ -56,13 +38,6 @@ namespace Prism
         {
             return new DebugLogger();
         }
-
-        /// <summary>
-        /// Creates the root <see cref="Xamarin.Forms.Page"/> for the application.
-        /// </summary>
-        /// <returns>The <see cref="Xamarin.Forms.Page"/></returns>
-        [Obsolete]
-        protected virtual Page CreateMainPage() => null;
 
         /// <summary>
         /// Creates the <see cref="INavigationService"/> for the application.
@@ -76,7 +51,7 @@ namespace Prism
         protected abstract void ConfigureViewModelLocator();
 
         /// <summary>
-        /// Called when the bootstrapper has completed it's initialization process.
+        /// Called when the PrismApplication has completed it's initialization process.
         /// </summary>
         protected abstract void OnInitialized();
