+++ Source/Xamarin/Prism.Ninject.Forms/Prism.Ninject.Forms.csproj
@@ -56,6 +56,7 @@
     <Compile Include="Extensions\DependencyServiceBindingResolver.cs" />
     <Compile Include="Extensions\DependencyServiceProvider.cs" />
     <Compile Include="Navigation\NinjectNavigationService.cs" />
+    <Compile Include="PrismApplication.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="NinjectBootstrapper.cs" />
     <Compile Include="NinjectExtensions.cs" />
